+++ src/main/java/com/tkl/refactoring/chapter1/StatementPrinter.java
@@ -1,51 +1,149 @@
 package com.tkl.refactoring.chapter1;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
 import java.text.NumberFormat;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
 public class StatementPrinter {
-
-    public String print(Invoice invoice, Map<String, Play> plays) {
-        int totalAmount = 0;
-        int volumeCredits = 0;
-        String result = String.format("Statement for %s\n", invoice.customer());
-
-        NumberFormat frmt = NumberFormat.getCurrencyInstance(Locale.US);
-
-        for (Performance perf : invoice.performances()) {
-            Play play = plays.get(perf.playID());
-            int thisAmount;
-
+    static class PerformanceCalculatorFactory {
+        public PerformanceCalculator createPerformanceCalculator(Performance performance, Play play) {
             switch (play.type()) {
                 case "tragedy":
-                    thisAmount = 40000;
-                    if (perf.audience() > 30) {
-                        thisAmount += 1000 * (perf.audience() - 30);
-                    }
-                    break;
+                    return new TragedyCalculator(performance, play);
                 case "comedy":
-                    thisAmount = 30000;
-                    if (perf.audience() > 20) {
-                        thisAmount += 10000 + 500 * (perf.audience() - 20);
-                    }
-                    thisAmount += 300 * perf.audience();
-                    break;
+                    return new ComedyCalculator(performance, play);
                 default:
-                    throw new Error("unknown type: ${play.type}");
+                    throw new Error("unknown type: ${playFor(performance).type()}");
             }
-
-            // add volume credits
-            volumeCredits += Math.max(perf.audience() - 30, 0);
-            // add extra credit for every ten comedy attendees
-            if ("comedy".equals(play.type())) volumeCredits += perf.audience() / 5;
-
-            // print line for this order
-            result += String.format("  %s: %s (%s seats)\n", play.name(), frmt.format(thisAmount / 100), perf.audience());
-            totalAmount += thisAmount;
         }
-        result += String.format("Amount owed is %s\n", frmt.format(totalAmount / 100));
-        result += String.format("You earned %s credits\n", volumeCredits);
+    }
+    @Getter
+    @AllArgsConstructor
+    static abstract class PerformanceCalculator {
+        protected Performance performance;
+        protected Play play;
+
+        protected abstract int amount();
+        protected abstract int volumeCredits();
+    }
+
+    static class TragedyCalculator extends PerformanceCalculator{
+        public TragedyCalculator(Performance performance, Play play) {
+            super(performance, play);
+        }
+        @Override
+        public int amount() {
+            int result = 40000;
+            if (performance.audience() > 30) {
+                result += 1000 * (performance.audience() - 30);
+            }
+            return result;
+        }
+
+        @Override
+        protected int volumeCredits() {
+            int result = 0;
+            result += Math.max(performance.audience() - 30, 0);
+            return result;
+        }
+    }
+    static class ComedyCalculator extends PerformanceCalculator{
+        public ComedyCalculator(Performance performance, Play play) {
+            super(performance, play);
+        }
+        @Override
+        public int amount() {
+            int result = 30000;
+            if (performance.audience() > 20) {
+                result += 10000 + 500 * (performance.audience() - 20);
+            }
+            result += 300 * performance.audience();
+            return result;
+        }
+
+        @Override
+        public int volumeCredits() {
+            int result = 0;
+            result += Math.max(performance.audience() - 30, 0);
+            result += performance.audience() / 5;
+            return result;
+        }
+    }
+
+    @Getter
+    static class StatementData {
+        private String customer;
+        private List<Performance> performances;
+        private Map<String, Play> plays;
+
+        public StatementData(Invoice invoice, Map<String, Play> plays) {
+            this.customer = invoice.customer();
+            this.performances = invoice.performances();
+            this.plays = plays;
+        }
+
+        public Play playFor(Performance performance) {
+            return plays.get(performance.playID());
+        }
+        private int amountFor(Performance performance) {
+            return new PerformanceCalculatorFactory().createPerformanceCalculator(performance, playFor(performance)).amount();
+        }
+        private int volumeCreditsFor(Performance performance) {
+            return new PerformanceCalculatorFactory().createPerformanceCalculator(performance, playFor(performance)).volumeCredits();
+//            return new PerformanceCalculator(performance, playFor(performance)).volumeCredits();
+        }
+        private int totalAmount(StatementData statementData) {
+            return statementData.performances.stream()
+                    .mapToInt(this::amountFor)
+                    .sum();
+        }
+        private int totalVolumeCredits(StatementData statementData) {
+            return statementData.getPerformances().stream()
+                    .mapToInt(this::volumeCreditsFor)
+                    .sum();
+        }
+    }
+
+    public String print(Invoice invoice, Map<String, Play> plays) {
+        return renderPlainText(createStatementData(invoice, plays));
+    }
+
+    private StatementData createStatementData(Invoice invoice, Map<String, Play> plays) {
+        return new StatementData(invoice, plays);
+    }
+
+    private String renderPlainText(StatementData statementData) {
+        StringBuilder result = new StringBuilder(String.format("Statement for %s\n", statementData.getCustomer()));
+
+        for (Performance performance : statementData.getPerformances()) {
+            result.append(String.format("  %s: %s (%s seats)\n", statementData.playFor(performance).name(), usd(statementData.amountFor(performance)), performance.audience()));
+        }
+
+        result.append(String.format("Amount owed is %s\n", usd(statementData.totalAmount(statementData))));
+        result.append(String.format("You earned %s credits\n", statementData.totalVolumeCredits(statementData)));
+        return result.toString();
+    }
+
+    private String usd(int amount) {
+        return NumberFormat.getCurrencyInstance(Locale.US).format(amount / 100);
+    }
+
+    private String renderHtml(StatementData statementData) {
+        String result = String.format("<h1>Statement for %s\n<h1>", statementData.getCustomer());
+        result += "<table>\n";
+        result += "<tr><th>연극</th><th><좌석 수></th><th>금액</th></tr>";
+
+        for (Performance performance : statementData.getPerformances()) {
+            result += String.format("  <tr><td>%s: </td><td>%s </td> <td>(%s seats)</td></tr>\n", statementData.playFor(performance).name(), usd(statementData.amountFor(performance)), performance.audience());
+        }
+        result += "</table>\n";
+
+        result += String.format("<p>Amount owed is <em> %s </em></p>\n", usd(statementData.totalAmount(statementData)));
+        result += String.format("<p>You earned <em> %s </em> credits </p>\n", statementData.totalVolumeCredits(statementData));
         return result;
     }
