--- src/test/java/com/tkl/refactoring/chapter1/StatementPrinterV1Test.java
@@ -1,54 +1,0 @@
-package com.tkl.refactoring.chapter1;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-
-class StatementPrinterV1Test {
-    static final String GIVEN_RESULT = """
-            Statement for BigCo
-              Hamlet: $650.00 (55 seats)
-              As You Like It: $580.00 (35 seats)
-              Othello: $500.00 (40 seats)
-            Amount owed is $1,730.00
-            You earned 47 credits
-            """;
-
-    @Test
-    void exampleStatement() {
-        Map<String, Play> plays = Map.of(
-                "hamlet", new Play("Hamlet", "tragedy"),
-                "as-like", new Play("As You Like It", "comedy"),
-                "othello", new Play("Othello", "tragedy"));
-
-        Invoice invoice = new Invoice("BigCo", List.of(
-                new Performance("hamlet", 55),
-                new Performance("as-like", 35),
-                new Performance("othello", 40)));
-
-        StatementPrinterV1 statementPrinter = new StatementPrinterV1();
-        var result = statementPrinter.print(invoice, plays);
-
-        assertThat(result).isEqualTo(GIVEN_RESULT);
-    }
-
-    @Test
-    void statementWithNewPlayTypes() {
-        Map<String, Play> plays = Map.of(
-                "henry-v", new Play("Henry V", "history"),
-                "as-like", new Play("As You Like It", "pastoral"));
-
-        Invoice invoice = new Invoice("BigCo", List.of(
-                new Performance("henry-v", 53),
-                new Performance("as-like", 55)));
-
-        StatementPrinterV1 statementPrinter = new StatementPrinterV1();
-        Assertions.assertThrows(Error.class, () -> statementPrinter.print(invoice, plays));
-    }
-
-}
